Rename and convert surfingkeys.js to surfingkeys-extended.js
This also makes sure the configuration works with a non-modified
version of the surfingkeys extension, which of course makes some
features worse. Namely the new tab binding does not focus the search
bar and the go to last/first tab bindings are slower.

Because of this you should use the provided smaller surfingkeys.js
configuration and config.js which can add bindings that are not
possible with modern web extensions.